refactor to use tokens pattern, clean up some signals
