+++ packages/jupyterlab-fonts/src/license.tsx
@@ -1,7 +1,7 @@
 import { VDomModel, VDomRenderer } from '@jupyterlab/apputils';
 import * as React from 'react';
 
-import { IFontFaceOptions } from '.';
+import { IFontFaceOptions } from './tokens';
 
 import '../style/license.css';
